requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
h5py>=3.9.0
einops>=0.7.0
transformers>=4.37.0
open_clip_torch>=2.24.0
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
wandb>=0.16.0
